refactor : CardList 타입 추가
